fix: 성하님 코드를 기존 코드와 병합 리펙토링
