+++ src/main/java/com/DrWait/domain/family/entity/FamilyGroup.java
@@ -1,9 +1,11 @@
 package com.DrWait.domain.family.entity;
 
+import com.DrWait.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
 
-import java.util.UUID;
+import java.util.HashSet;
+import java.util.Set;
 
 @Entity
 @Getter @Setter
@@ -17,5 +19,28 @@ public class FamilyGroup {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private UUID userId;
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User owner;
+
+    @Getter
+    @OneToMany(mappedBy = "familyGroup",
+            cascade = CascadeType.ALL,
+            orphanRemoval = true,
+            fetch = FetchType.LAZY)
+    private Set<FamilyMember> members = new HashSet<>();
+
+    public FamilyGroup(User owner) {
+        this.owner = owner;
+    }
+
+    public void addMember(FamilyMember member) {
+        members.add(member);
+        member.setFamilyGroup(this);
+    }
+
+    public void removeMember(FamilyMember member) {
+        members.remove(member);
+        member.setFamilyGroup(null);
+    }
 }
