+++ src/main/java/com/DrWait/domain/family/entity/FamilyMember.java
@@ -1,10 +1,8 @@
 package com.DrWait.domain.family.entity;
 
-import com.DrWait.domain.family.enums.FamilyRole;
+import com.DrWait.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
-
-import java.util.UUID;
 
 @Entity
 @Getter @Setter
@@ -14,11 +12,24 @@ import java.util.UUID;
 @Table(name = "family_member")
 public class FamilyMember {
 
-    @EmbeddedId
+    @EmbeddedId // Composite primary key
     private FamilyMemberId id;
 
-    @Enumerated(EnumType.STRING)
-    private FamilyRole role;
+    @MapsId("familyGroupId")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "family_group_id", nullable = false)
+    private FamilyGroup familyGroup;
 
-    private boolean isConfirmed;
+    @MapsId("userId")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    @Column(nullable = false, length = 255)
+    private String role;
+
+    @Column(name = "is_confirmed",
+            nullable = false,
+            columnDefinition = "TINYINT(1) DEFAULT 0")
+    private boolean isConfirmed = false;
 }
