--- src/main/java/com/DrWait/domain/family/enums/FamilyRole.java
@@ -1,7 +1,0 @@
-package com.DrWait.domain.family.enums;
-
-public enum FamilyRole {
-    OWNER,  // 대표자 1명
-    MEMBER, // 일반 가족 구성원
-    GUEST   // 초대받았지만 미확정된 사용자 등
-}
