+++ src/main/java/com/DrWait/domain/family/repository/FamilyGroupRepository.java
@@ -1,6 +1,7 @@
 package com.DrWait.domain.family.repository;
 
 import com.DrWait.domain.family.entity.FamilyGroup;
+import com.DrWait.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,6 +9,5 @@ import java.util.Optional;
 
 @Repository
 public interface FamilyGroupRepository extends JpaRepository<FamilyGroup, Long> {
-
-    Optional<FamilyGroup> findById(Long id);
+    Optional<FamilyGroup> findByOwner(User owner);
 }
