+++ src/main/java/com/DrWait/domain/family/repository/FamilyMemberRepository.java
@@ -1,6 +1,7 @@
 package com.DrWait.domain.family.repository;
 
 import com.DrWait.domain.family.entity.FamilyMember;
+import com.DrWait.domain.family.entity.FamilyMemberId;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,9 +9,5 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Repository
-public interface FamilyMemberRepository extends JpaRepository<FamilyMember, Long> {
-
-    Optional<FamilyMember> findByUserId(UUID userId);
-
-    Boolean existsById_UserId(UUID userId);
+public interface FamilyMemberRepository extends JpaRepository<FamilyMember, FamilyMemberId> {
 }
